api/requirements-lite.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
PyYAML==6.0.1

# Deep analysis dependencies
openai==1.12.0
yt-dlp==2024.10.7
ffmpeg-python==0.2.0

